requirements.txt
playfab
python-dotenv